+++ src/util.c
@@ -1,8 +1,8 @@
 #include "util.h"
 
-unsigned int current_time;
+unsigned long long current_time;
 
-unsigned int get_current_time() {
+unsigned long long get_current_time() {
     return current_time;
 }
